refactor(models): Se cambio la ruta del repositorio django y se reestablecio steps
